requirements.txt
lxml[html_clean]
transformers
langchain
newspaper3k
feedparser
langchain_community
torch
langdetect
